+++ source/elegant/src/pages/docs/[slug]/index.tsx
@@ -1,24 +1,22 @@
 import { getDocumentBySlug } from "@/utils/core/Collections/collection";
 import { GetServerSideProps } from "next";
-import { SidebarLayout } from "@/layouts/SidebarLayout";
 import Error from "@/pages/404";
 import MarkdownToHtml from "@/utils/core/Rehype/MarkdownToHtml";
 import { useRouter } from "next/router";
-import { documentationNav } from '@/config/Navigation';
-import Head from 'next/head';
-import AnalyticsHead from '@/components/core/Analytics/AnalyticsHead';
-import Config from 'Config';
-import Seo from "@/components/core/Seo/Seo";
-import socialCardLarge from '@/img/social-card-large.jpg';
-import MetaTitle from "@/utils/core/Meta/MetaTitle";
 import HtmlToToc from "@/utils/core/Rehype/HtmlToToc";
 import DocumentationLayout from "@/components/core/Layouts/DocumentationLayout";
 import { Post } from "@/types/Post";
-import SocialSchema from "@/utils/core/Meta/SocialSchema";
-import { BrandJsonLd, LogoJsonLd, SocialProfileJsonLd, WebPageJsonLd } from "next-seo";
-import SeoLogo from './../../public/favicons/apple-icon-180x180.png';
-import { useEffect, useState } from "react";
-import * as gtag from '@/utils/core/Analytics/gtag';
+import { DocumentationHeading } from "@/components/core/Headings/DocumentationHeading";
+import TableOfContents from "@/components/core/TableOfContents/TableOfContents";
+import { TableOfContentsItem } from "@/types/TableOfContentsItem";
+import { createContext, useCallback, useEffect, useState } from "react";
+import { usePrevNext } from "@/utils/core/Hooks/usePrevNext";
+import { Heading } from "@/types/Heading";
+import DocsFooter from "@/components/core/Footer/DocsFooter/DocsFooter";
+import Config from "@/utils/core/Config/Config";
+import Link from "next/link";
+
+export const ContentsContext = createContext({});
 
 type Props = {
     /**
@@ -35,6 +33,14 @@ type Props = {
     toc: [];
 };
 
+Index.layoutProps = {
+    Layout: DocumentationLayout,
+    meta: {
+        title: "Hey",
+        section: "Now"
+    }
+}
+
 /**
  * Our documentation page that dynamically loads .mdx files based on our url slug
  * @returns An html page for displaying documentation.
@@ -45,39 +51,6 @@ export default function Index({
     toc
 }: Props) {
     const router = useRouter();
-    let [navIsOpen, setNavIsOpen] = useState(false);
-
-    useEffect(() => {
-        if (!navIsOpen) {
-            return;
-        }
-
-        function handleRouteChange() {
-          setNavIsOpen(false);
-        }
-
-        router.events.on('routeChangeComplete', handleRouteChange);
-
-        return () => {
-            router.events.off('routeChangeComplete', handleRouteChange);
-        }
-      }, [navIsOpen]);
-
-    /**
-   * Fire a page view analytics event when a user navigates to a new page
-   */
-  useEffect(() => {
-    if(Config('app.google_analytics_id').length > 0){
-      const handleRouteChange = (url: string) => {
-        gtag.pageview(url);
-      }
-      router.events.on('routeChangeComplete', handleRouteChange);
-
-      return () => {
-        router.events.off('routeChangeComplete', handleRouteChange);
-      }
-    }
-  }, [router.events]);
 
     if (!router.isFallback && !post?.slug) {
         return (
@@ -85,83 +58,45 @@ export default function Index({
         );
     }
 
-    // set our url
-    let url = Config('app.url');
-
-    if(
-        process.env.NEXT_PUBLIC_VERCEL_URL !== undefined &&
-        process.env.NEXT_PUBLIC_VERCEL_ENV !== undefined &&
-        process.env.NEXT_PUBLIC_VERCEL_ENV !== "production"
-    ){
-        url = "https://" + process.env.NEXT_PUBLIC_VERCEL_URL;
-    }
-
-    // if there are social links provided, activate the Social schema data
-    let socialSchema = SocialSchema();
-
-    // Set the social share image
-    let image = socialCardLarge.src;
-
-    if(post.coverImage){
-        image = post.coverImage;
-    }
+    const { currentSection, registerHeading, unregisterHeading } = useTableOfContents(toc);
+    let { prev, next } = usePrevNext();
 
     return(
-        <>
-            <Head>
-                <AnalyticsHead googleAnalyticsID={Config('app.google_analytics_id')}/>
-            </Head>
-            <Seo 
-                title={`${post.title} - ${MetaTitle(Config('app.name'), Config('app.tagline'))}`}
-                description={post.description || Config('app.description')}
-                themeColor={"#f8fafc"}
-                twitterHandle={Config('app.twitter_handle')}
-                twitterSite={Config('app.twitter_handle')}
-                siteName={Config('app.name')}
-                url={`${url}${router.pathname}`}
-                image={`${url}${image}`}
-                facebookAppID={Config('app.facebook_app_id')}
-                pageType={'website'}
-            />
-            <LogoJsonLd 
-                logo={Config('app.url') + SeoLogo.src}
-                url={Config('app.url')}
-            />
-            <BrandJsonLd 
-                logo={Config('app.url') + SeoLogo.src}
-                id={Config('app.url')}
-                slogan={Config('app.tagline')}
-            />
-            <WebPageJsonLd 
+        <div className="max-w-3xl mx-auto pt-10 xl:max-w-none xl:ml-0 xl:mr-[15.5rem] xl:pr-16">
+            <DocumentationHeading 
+                title={post.title}
                 description={post.description}
-                id={`${Config('app.url')}${router.pathname}`}
+                section={post.section}
             />
-            {socialSchema.length > 0 && (
-                <SocialProfileJsonLd 
-                    type={
-                        (Config('app.type') === "Person" ? "Person" : "Organization")
+            <ContentsContext.Provider value={{ registerHeading, unregisterHeading }}>
+                <div
+                    id="content-wrapper"
+                    className="relative z-20 prose prose-slate mt-8 dark:prose-dark"
+                >
+                    <div dangerouslySetInnerHTML={{ __html: content }} />
+                </div>
+                <DocsFooter previous={prev} next={next}>
+                    {Config('app.repository').length > 0 &&
+                        <Link
+                            href={`${Config('app.repository')}/edit/main/${Config('admin.cms_monorepo_path')}${Config('admin.cms_content_path')}/docs/${post.slug}.mdx`}
+                            className="hover:text-slate-900 dark:hover:text-slate-400"
+                            target="_blank"
+                            rel="noopener noreferrer"
+                        >
+                            Edit this page on GitHub
+                        </Link>
                     }
-                    name={Config('app.name')}
-                    url={url}
-                    sameAs={socialSchema}
-                />
-            )}
-            <SidebarLayout
-                navIsOpen={navIsOpen}
-                setNavIsOpen={
-                    (isOpen: boolean) => setNavIsOpen(isOpen)
-                }
-                nav={documentationNav}
-                sidebar={null}
-                fallbackHref={null}
-            >
-                <DocumentationLayout 
-                    post={post}
-                    content={content}
-                    toc={toc}
-                />
-            </SidebarLayout>
-        </>
+                </DocsFooter>
+                <div className="fixed z-20 top-[3.8125rem] bottom-0 right-[max(0px,calc(50%-45rem))] w-[19.5rem] py-10 overflow-y-auto hidden xl:block">
+                    {toc.length > 0 && (
+                        <TableOfContents 
+                            tableOfContents={toc} 
+                            currentSection={currentSection} 
+                        />
+                    )}
+                </div>
+            </ContentsContext.Provider>
+        </div>
     );
 };
 
@@ -189,3 +124,75 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
         }
     };
 };
+
+/**
+ * 
+ * @param tableOfContents 
+ * @returns 
+ */
+function useTableOfContents(tableOfContents: TableOfContentsItem[]) {
+
+    let [currentSection, setCurrentSection] = useState(tableOfContents[0]?.slug)
+    let [headings, setHeadings] = useState<Heading[]>([])
+  
+    const registerHeading = useCallback((id: string) => {
+        setHeadings((headings) => [...headings.filter((h) => id !== h.id), { id, top: getTop(id) }])
+    }, [])
+  
+    const unregisterHeading = useCallback((id: string) => {
+        setHeadings((headings) => headings.filter((h) => id !== h.id))
+    }, [])
+  
+    useEffect(() => {
+      if (tableOfContents.length === 0 || headings.length === 0) {
+        return;
+      }
+  
+      function onScroll() {
+        let style = window.getComputedStyle(document.documentElement)
+        let scrollMt = parseFloat(style.getPropertyValue('--scroll-mt').match(/[\d.]+/)?.[0] as string ?? 0)
+        let fontSize = parseFloat(style.fontSize.match(/[\d.]+/)?.[0] as string ?? 16)
+        scrollMt = scrollMt * fontSize;
+        
+        let sortedHeadings = headings.concat([]).sort((a, b) => a.top - b.top)
+        let top = window.pageYOffset + scrollMt + 1
+        let current = sortedHeadings[0].id
+        for (let i = 0; i < sortedHeadings.length; i++) {
+          if (top >= sortedHeadings[i].top) {
+            current = sortedHeadings[i].id
+          }
+        }
+        setCurrentSection(current)
+      }
+  
+      window.addEventListener('scroll', onScroll, {
+        capture: true,
+        passive: true,
+      })
+  
+      onScroll()
+  
+      let resizeObserver = new window.ResizeObserver(() => {
+        for (let heading of headings) {
+          heading.top = getTop(heading.id)
+        }
+      })
+  
+      resizeObserver.observe(document.body)
+      return () => {
+        resizeObserver.disconnect()
+        window.removeEventListener('scroll', onScroll, {
+          capture: true,
+          //@ts-ignore
+          passive: true,
+        })
+      }
+    }, [headings, tableOfContents])
+  
+    return { currentSection, registerHeading, unregisterHeading }
+  }
+
+function getTop(id: string) {
+    //@ts-ignore
+    return document.getElementById(id).getBoundingClientRect().top + window.scrollY;
+}
